* Renames: Personalisation to Personalization package
