+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/personalization/DashboardCardState.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.mysite.personalisation
+package org.wordpress.android.ui.mysite.personalization
 
 import androidx.annotation.StringRes
